Moved ErrorMessage enum to types folder
