+++ src/features/CourtCaseDetails/Tabs/Panels/Offences/Offence/OffenceDetails.tsx
@@ -2,11 +2,11 @@ import { Offence } from "@moj-bichard7-developers/bichard7-next-core/core/types/
 import { ExceptionCode } from "@moj-bichard7-developers/bichard7-next-core/core/types/ExceptionCode"
 import offenceCategory from "@moj-bichard7-developers/bichard7-next-data/dist/data/offence-category.json"
 import yesNo from "@moj-bichard7-developers/bichard7-next-data/dist/data/yes-no.json"
-import ErrorMessages from "Data/ErrorMessages"
 import SecondaryButton from "components/SecondaryButton"
 import UneditableField from "components/UneditableField"
 import { GridCol, GridRow, Heading, Table } from "govuk-react"
 import { createUseStyles } from "react-jss"
+import ErrorMessages from "types/ErrorMessages"
 import { formatDisplayedDate } from "utils/formattedDate"
 import getOffenceCode from "utils/getOffenceCode"
 import { TableRow } from "../../TableRow"
